requirements.txt
faster-whisper
transformers
ffmpeg
ffmpeg-python
pycairo
PyGObject
regex
fuzzysearch
python-magic
pandas
torch
torchaudio
auditok
stable-ts==2.17.3
retry2
num2words@git+https://github.com/patriotyk/num2words.git

